Some progress
Materials, sphere ray intersection, world as a container of objects

+++ Ray.cpp
@@ -1,6 +1,6 @@
 #include "Ray.hpp"
 
-Vec3 Ray::at(float t) const
+Vec3 Ray::at(double t) const
 {
 	return Origin + t * Direction;
 }
